#110 change broadcasting of constant vector to be consistent with StateVector
